Support relative and absolute binary paths
Our first bug report! This was something the `which` crate handled for us: passing in an absolute path to the command to run used to skip the $PATH search, but our own `which` implementation didn’t check for that. The code is a bit longer as a result, but I think this is an expected feature.