Hide the existence of non-executable files
This commit changes which files are returned by the PATH scanner — it now no longer returns non-executable files. This deals with a problem where a user could detect the existence of a file by looking at the error message when trying to run it through minisudo (which was CVE-1999-1496 for sudo).

The problem still exists for executable files.